backend/requirements.txt
Django==4.2.14
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-redis>=5.4.0
dj-database-url==3.0.1
django-environ==0.12.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg[binary]==3.2.9
Pillow==11.3.0
python-dotenv==1.1.1
# ↑ dotenv と decouple はどちらか一方に統一推奨。decoupleを使うなら下行を有効化してdotenvを削除。
# python-decouple==3.8
djangorestframework-simplejwt>=5.3
requests>=2.31.0

python-dotenv>=1.0.0
urllib3<2
